pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "mazurkevych30",email = "132198562+mazurkevych30@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "apscheduler (>=3.11.0,<4.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "cloudinary (>=1.44.0,<2.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-asyncio = "^0.26.0"
aiosqlite = "^0.21.0"
pytest-cov = "^6.1.1"
sphinx = "^8.2.3"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--doctest-modules -W ignore::DeprecationWarning -W ignore::UserWarning"
testpaths = ["tests"]
pythonpath = "."
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning"
]
asyncio_default_fixture_loop_scope = "function"




